implement individual file removal
adds an icon to each file in the file list that indicates to the user that that particular file can be deleted. When the file is deleted the preprocessed filesize is recalculated. This was added in case adds an image they did not mean to, or if they add an invalid file format. This was the user doesn't have to re-add all the files, they can just remove the offending file.